WIP truncation
../lib/gen_tables.c:89:22: warning: overflow in conversion from ‘long long unsigned int’ to ‘int’ changes value from ‘4294967296’ to ‘0’ [-Woverflow]
   89 | #define _S(VAL, S) { (VAL), (S), 0, 0 },
      |                      ^
./clone-flagtab.h:47:1: note: in expansion of macro ‘_S’
   47 | _S(0x100000000ULL,      "CLONE_CLEAR_SIGHAND")
      | ^~

+++ audisp/plugins/syslog/audisp-syslog.c
@@ -158,7 +158,7 @@ static inline void write_syslog(char *s)
 		mptr = record;
 		while (rc > 0 &&
 		       ((mptr-record) < (MAX_AUDIT_MESSAGE_LENGTH-128))) {
-			int ftype = auparse_get_field_type(au);
+			long long int ftype = auparse_get_field_type(au);
 			const char *fname = auparse_get_field_name(au);
 			const char *fval;
 			switch (ftype) {
